Add commit-changes to battle interface and implement
Battle now has two event accumulators, one for incoming events and one
for outgoing events. Commit-changes processes the events in the
incoming queue and transfers them to the outgoing queue.

To make a test pass, I had to implement perform-event for the death
event, but it doesn't actually do anything yet.